Updates from homebrew-realcore 20211130.213738

+++ Formula/pumba.rb
@@ -1,8 +1,8 @@
 class Pumba < Formula
   desc "Chaos testing tool for Docker"
   homepage "https://github.com/alexei-led/pumba"
-  url "https://github.com/alexei-led/pumba/archive/0.8.0.tar.gz"
-  sha256 "052ece6984a0533d7f93b2b64c66d5e89516bbf93e4cb732a2743322b4eef9da"
+  url "https://github.com/alexei-led/pumba/archive/0.9.0.tar.gz"
+  sha256 "7faa50566898a53b0fff81973e7161874eabec45ad11f9defcd0e04310bddaff"
   license "Apache-2.0"
   head "https://github.com/alexei-led/pumba.git", branch: "master"
 
@@ -11,20 +11,22 @@ class Pumba < Formula
     strategy :github_latest
   end
 
-  bottle do
-    root_url "https://github.com/gromgit/homebrew-core-mojave/releases/download/pumba"
-    rebuild 1
-    sha256 cellar: :any_skip_relocation, mojave: "441c705b47ee57fda3e27a2c6df00c9c79e7f8ae221a29f6bba51d60fcee7cc7"
-  end
-
   depends_on "go" => :build
 
   def install
-    system "go", "build", "-ldflags", "-s -w -X main.Version=#{version}",
-           "-trimpath", "-o", bin/"pumba", "./cmd"
+    goldflags = %W[
+      -s -w
+      -X main.Version=#{version}
+      -X main.BuildTime=#{time.iso8601}
+    ].join(" ")
+    system "go", "build", *std_go_args(ldflags: goldflags), "./cmd"
   end
 
   test do
+    assert_match version.to_s, shell_output("#{bin}/pumba --version")
+    # CI runs in a Docker container, so the test does not run as expected.
+    return if OS.linux? && ENV["HOMEBREW_GITHUB_ACTIONS"].present?
+
     output = pipe_output("#{bin}/pumba rm test-container 2>&1")
     assert_match "Is the docker daemon running?", output
   end
